+++ src/main/java/gregtech/api/metatileentity/multiblock/MultiblockAbility.java
@@ -5,12 +5,14 @@ import gregtech.api.capability.IMaintenanceHatch;
 import gregtech.api.capability.IMufflerHatch;
 import gregtech.api.capability.IRotorHolder;
 import gregtech.api.metatileentity.MetaTileEntity;
-import gregtech.common.metatileentities.multi.multiblockpart.MetaTileEntityRotorHolder;
 import it.unimi.dsi.fastutil.objects.Object2ObjectOpenHashMap;
 import net.minecraftforge.fluids.IFluidTank;
 import net.minecraftforge.items.IItemHandlerModifiable;
 
-import java.util.*;
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
 
 @SuppressWarnings("InstantiationOfUtilityClass")
 public class MultiblockAbility<T> {
@@ -26,9 +28,7 @@ public class MultiblockAbility<T> {
     public static final MultiblockAbility<IEnergyContainer> INPUT_ENERGY = new MultiblockAbility<>("input_energy");
     public static final MultiblockAbility<IEnergyContainer> OUTPUT_ENERGY = new MultiblockAbility<>("output_energy");
 
-    public static final MultiblockAbility<MetaTileEntityRotorHolder> ABILITY_ROTOR_HOLDER = new MultiblockAbility<>("ability_rotor_holder");
-
-    public static final MultiblockAbility<IRotorHolder> ABILITY_ROTOR_HOLDER_2 = new MultiblockAbility<>("rotor_holder");
+    public static final MultiblockAbility<IRotorHolder> ROTOR_HOLDER = new MultiblockAbility<>("rotor_holder");
 
     public static final MultiblockAbility<IFluidTank> PUMP_FLUID_HATCH = new MultiblockAbility<>("pump_fluid_hatch");
