+++ src/main/java/gregtech/common/metatileentities/multi/electric/generator/LargeTurbineWorkableHandler.java
@@ -1,153 +1,150 @@
 package gregtech.common.metatileentities.multi.electric.generator;
 
-import gregtech.api.capability.IEnergyContainer;
-import gregtech.api.capability.IMultipleTankHandler;
-import gregtech.api.capability.impl.FuelRecipeLogic;
-import gregtech.api.metatileentity.multiblock.MultiblockAbility;
-import gregtech.api.recipes.machines.FuelRecipeMap;
-import gregtech.api.recipes.recipes.FuelRecipe;
-import gregtech.api.unification.material.Material;
-import gregtech.api.unification.material.Materials;
-import gregtech.common.ConfigHolder;
-import gregtech.common.MetaFluids;
-import gregtech.common.metatileentities.multi.multiblockpart.MetaTileEntityRotorHolder;
-import gregtech.common.metatileentities.multi.electric.generator.MetaTileEntityLargeTurbine.TurbineType;
-import net.minecraft.nbt.NBTTagCompound;
+import gregtech.api.GTValues;
+import gregtech.api.capability.IRotorHolder;
+import gregtech.api.capability.impl.MultiblockFuelRecipeLogic;
+import gregtech.api.metatileentity.multiblock.RecipeMapMultiblockController;
+import gregtech.api.recipes.Recipe;
+import gregtech.api.recipes.RecipeBuilder;
 import net.minecraft.util.math.MathHelper;
 import net.minecraftforge.fluids.FluidStack;
 
-import java.util.function.Supplier;
+public class LargeTurbineWorkableHandler extends MultiblockFuelRecipeLogic {
 
-public class LargeTurbineWorkableHandler extends FuelRecipeLogic {
+    private final int BASE_EU_OUTPUT;
 
-    private static final int CYCLE_LENGTH = 230;
-    private static final int BASE_ROTOR_DAMAGE = 11;
-    private static final int BASE_EU_OUTPUT = 2048;
+    private int excessVoltage;
 
-    private final MetaTileEntityLargeTurbine largeTurbine;
-    private int rotorCycleLength = CYCLE_LENGTH;
+    private boolean shouldRun = true;
 
-    public LargeTurbineWorkableHandler(MetaTileEntityLargeTurbine metaTileEntity, FuelRecipeMap recipeMap, Supplier<IEnergyContainer> energyContainer, Supplier<IMultipleTankHandler> fluidTank) {
-        super(metaTileEntity, recipeMap, energyContainer, fluidTank, 0L);
-        this.largeTurbine = metaTileEntity;
+    public LargeTurbineWorkableHandler(RecipeMapMultiblockController metaTileEntity, int tier) {
+        super(metaTileEntity);
+        this.BASE_EU_OUTPUT = (int) GTValues.V[tier] * 2;
     }
 
     @Override
     public void update() {
         super.update();
-        MetaTileEntityRotorHolder rotorHolder = largeTurbine.getAbilities(MultiblockAbility.ABILITY_ROTOR_HOLDER).get(0);
-        if (!rotorHolder.isHasRotor()) {
-            setActive(false);
-        }
-        long totalEnergyOutput = getRecipeOutputVoltage();
-        if (totalEnergyOutput > 0) {
-            energyContainer.get().addEnergy(totalEnergyOutput);
+        //update the status of the workable twice per second to reduce the density of checks
+        if (!metaTileEntity.getWorld().isRemote && metaTileEntity.getOffsetTimer() % 10 == 0) {
+            shouldRun = shouldRunUpdate();
         }
     }
 
-    public FluidStack getFuelStack() {
+    public FluidStack getInputFluidStack() {
         if (previousRecipe == null)
             return null;
-        FluidStack fuelStack = previousRecipe.getRecipeFluid();
-        return fluidTank.get().drain(new FluidStack(fuelStack.getFluid(), Integer.MAX_VALUE), false);
+        FluidStack fuelStack = previousRecipe.getFluidInputs().get(0);
+        return getInputTank().drain(new FluidStack(fuelStack.getFluid(), Integer.MAX_VALUE), false);
+    }
+
+    public boolean shouldRun() {
+        return shouldRun;
+    }
+
+    public boolean shouldRunUpdate() {
+        IRotorHolder rotorHolder = ((MetaTileEntityLargeTurbine) metaTileEntity).getRotorHolder();
+        return rotorHolder != null && rotorHolder.hasRotor() && rotorHolder.isFrontFaceFree();
     }
 
     @Override
-    public boolean checkRecipe(FuelRecipe recipe) {
-        MetaTileEntityRotorHolder rotorHolder = largeTurbine.getAbilities(MultiblockAbility.ABILITY_ROTOR_HOLDER).get(0);
-        if (++rotorCycleLength >= CYCLE_LENGTH) {
-            int damageToBeApplied = (int) Math.round(BASE_ROTOR_DAMAGE * rotorHolder.getRelativeRotorSpeed()) + 1;
-            if (rotorHolder.applyDamageToRotor(damageToBeApplied, false)) {
-                this.rotorCycleLength = 0;
-                return true;
-            } else return false;
+    public void updateRecipeProgress() {
+        //no safe void for turbines
+        drawEnergy(recipeEUt, false);
+        if (++progressTime > maxProgressTime) {
+            completeRecipe();
         }
-        return true;
+    }
+
+    @Override
+    protected void trySearchNewRecipe() {
+        if (shouldRun())
+            super.trySearchNewRecipe();
     }
 
     @Override
     public long getMaxVoltage() {
-        MetaTileEntityRotorHolder rotorHolder = largeTurbine.getAbilities(MultiblockAbility.ABILITY_ROTOR_HOLDER).get(0);
-        if (rotorHolder.hasRotorInInventory()) {
-            double rotorEfficiency = rotorHolder.getRotorEfficiency();
-            double totalEnergyOutput = (BASE_EU_OUTPUT + getBonusForTurbineType(largeTurbine) * rotorEfficiency);
-            return MathHelper.ceil(totalEnergyOutput);
+        IRotorHolder rotorHolder = ((MetaTileEntityLargeTurbine) metaTileEntity).getRotorHolder();
+        if (rotorHolder != null && rotorHolder.hasRotor())
+            return (long) BASE_EU_OUTPUT * rotorHolder.getTotalPower() / 100;
+        return 0;
+    }
+
+    @Override
+    protected long boostProduction(int production) {
+        IRotorHolder rotorHolder = ((MetaTileEntityLargeTurbine) metaTileEntity).getRotorHolder();
+        if (rotorHolder != null && rotorHolder.hasRotor()) {
+            int maxSpeed = rotorHolder.getMaxRotorHolderSpeed();
+            int currentSpeed = rotorHolder.getRotorSpeed();
+            if (currentSpeed >= maxSpeed)
+                return production;
+            return (int) (production * Math.pow(1.0 * currentSpeed / maxSpeed, 2));
         }
-        return BASE_EU_OUTPUT + getBonusForTurbineType(largeTurbine);
+        return 0;
     }
 
     @Override
-    protected boolean isReadyForRecipes() {
-        MetaTileEntityRotorHolder rotorHolder = largeTurbine.getAbilities(MultiblockAbility.ABILITY_ROTOR_HOLDER).get(0);
-        return rotorHolder.isHasRotor() && super.isReadyForRecipes();
-    }
+    protected boolean prepareRecipe(Recipe recipe) {
+        IRotorHolder rotorHolder = ((MetaTileEntityLargeTurbine) metaTileEntity).getRotorHolder();
+        if (rotorHolder == null || !rotorHolder.hasRotor())
+            return false;
 
-    @Override
-    protected long startRecipe(FuelRecipe currentRecipe, int fuelAmountUsed, int recipeDuration) {
-        addOutputFluids(currentRecipe, fuelAmountUsed);
-        return 0L; //energy is added each tick while the rotor speed is >0 RPM
-    }
+        int turbineMaxVoltage = (int) getMaxVoltage();
+        FluidStack recipeFluidStack = recipe.getFluidInputs().get(0);
+        int inputAmount = 0;
+        int parallel = 0;
 
-    private void addOutputFluids(FuelRecipe currentRecipe, int fuelAmountUsed) {
-        if (largeTurbine.turbineType == TurbineType.STEAM) {
-            int waterFluidAmount = fuelAmountUsed / 15;
-            if (waterFluidAmount > 0) {
-                FluidStack waterStack = Materials.Water.getFluid(waterFluidAmount);
-                largeTurbine.exportFluidHandler.fill(waterStack, true);
-            }
-        } else if (largeTurbine.turbineType == TurbineType.PLASMA) {
-            Material material = MetaFluids.getMaterialFromFluid(currentRecipe.getRecipeFluid().getFluid());
-            if (material != null) {
-                largeTurbine.exportFluidHandler.fill(material.getFluid(fuelAmountUsed), true);
-            }
+        if (excessVoltage >= turbineMaxVoltage) {
+            excessVoltage -= turbineMaxVoltage;
+        } else {
+            double holderEfficiency = rotorHolder.getTotalEfficiency() / 100.0;
+            //get the amount of parallel required to match the desired output voltage
+            parallel = MathHelper.ceil((turbineMaxVoltage - excessVoltage) /
+                            (Math.abs(recipe.getEUt()) * holderEfficiency));
+
+            inputAmount = recipeFluidStack.amount * parallel;
+            if (getInputFluidStack().amount < inputAmount)
+                return false;
+
+            excessVoltage += (int) (parallel * Math.abs(recipe.getEUt()) * holderEfficiency - turbineMaxVoltage);
         }
-    }
 
-    private int getBonusForTurbineType(MetaTileEntityLargeTurbine turbine) {
-        switch (turbine.turbineType) {
-            case GAS:
-                return ConfigHolder.machines.gasTurbineBonusOutput;
-            case PLASMA:
-                return ConfigHolder.machines.plasmaTurbineBonusOutput;
-            case STEAM:
-                return ConfigHolder.machines.steamTurbineBonusOutput;
-            default:
-                return 1;
+        //rebuild the recipe and adjust voltage to match the turbine, with adjusted input
+        RecipeBuilder<?> recipeBuilder = getRecipeMap().recipeBuilder();
+        recipeBuilder.append(recipe, parallel, false, false)
+                .EUt(-turbineMaxVoltage);
+        if (recipeFluidStack.amount != inputAmount) {
+            recipeBuilder.clearFluidInputs()
+                    .fluidInputs(new FluidStack(recipeFluidStack.getFluid(), inputAmount));
         }
-    }
+        recipe = recipeBuilder.build().getResult();
 
-    @Override
-    public long getRecipeOutputVoltage() {
-        MetaTileEntityRotorHolder rotorHolder = largeTurbine.getAbilities(MultiblockAbility.ABILITY_ROTOR_HOLDER).get(0);
-        double relativeRotorSpeed = rotorHolder.getRelativeRotorSpeed();
-        if (rotorHolder.getCurrentRotorSpeed() > 0 && rotorHolder.hasRotorInInventory()) {
-            double rotorEfficiency = rotorHolder.getRotorEfficiency();
-            double totalEnergyOutput = ((BASE_EU_OUTPUT + getBonusForTurbineType(largeTurbine)) * rotorEfficiency) * (relativeRotorSpeed * relativeRotorSpeed);
-            return MathHelper.ceil(totalEnergyOutput);
+        if (recipe != null && setupAndConsumeRecipeInputs(recipe, getInputInventory())) {
+            setupRecipe(recipe);
+            return true;
         }
-        return 0L;
+        return false;
     }
 
     @Override
-    public NBTTagCompound serializeNBT() {
-        NBTTagCompound tagCompound = super.serializeNBT();
-        tagCompound.setInteger("CycleLength", rotorCycleLength);
-        return tagCompound;
+    public void invalidate() {
+        excessVoltage = 0;
+        super.invalidate();
     }
 
-    @Override
-    public void deserializeNBT(NBTTagCompound compound) {
-        super.deserializeNBT(compound);
-        this.rotorCycleLength = compound.getInteger("CycleLength");
-    }
-
-    @Override
-    protected boolean shouldVoidExcessiveEnergy() {
-        return true;
-    }
-
-    @Override
-    protected boolean isObstructed() {
-        return !largeTurbine.isRotorFaceFree();
-    }
+    //TODO re-implement large turbine fluid output using recipe voiding
+//    private void addOutputFluids(FuelRecipe currentRecipe, int fuelAmountUsed) {
+//        if (largeTurbine.turbineType == TurbineType.STEAM) {
+//            int waterFluidAmount = fuelAmountUsed / 15;
+//            if (waterFluidAmount > 0) {
+//                FluidStack waterStack = Materials.Water.getFluid(waterFluidAmount);
+//                largeTurbine.exportFluidHandler.fill(waterStack, true);
+//            }
+//        } else if (largeTurbine.turbineType == TurbineType.PLASMA) {
+//            Material material = MetaFluids.getMaterialFromFluid(currentRecipe.getRecipeFluid().getFluid());
+//            if (material != null) {
+//                largeTurbine.exportFluidHandler.fill(material.getFluid(fuelAmountUsed), true);
+//            }
+//        }
+//    }
 }
