+++ src/main/java/gregtech/common/metatileentities/multi/electric/generator/MetaTileEntityLargeTurbine.java
@@ -1,23 +1,20 @@
 package gregtech.common.metatileentities.multi.electric.generator;
 
 import gregtech.api.GTValues;
+import gregtech.api.capability.IRotorHolder;
 import gregtech.api.capability.impl.FluidTankList;
-import gregtech.api.capability.impl.FuelRecipeLogic;
+import gregtech.api.metatileentity.ITieredMetaTileEntity;
 import gregtech.api.metatileentity.MetaTileEntity;
 import gregtech.api.metatileentity.MetaTileEntityHolder;
+import gregtech.api.metatileentity.multiblock.FuelMultiblockController;
 import gregtech.api.metatileentity.multiblock.IMultiblockPart;
 import gregtech.api.metatileentity.multiblock.MultiblockAbility;
 import gregtech.api.pattern.BlockPattern;
 import gregtech.api.pattern.FactoryBlockPattern;
 import gregtech.api.pattern.PatternMatchContext;
 import gregtech.api.pattern.TraceabilityPredicate;
-import gregtech.api.recipes.RecipeMaps;
-import gregtech.api.recipes.machines.FuelRecipeMap;
+import gregtech.api.recipes.RecipeMap;
 import gregtech.client.renderer.ICubeRenderer;
-import gregtech.client.renderer.texture.Textures;
-import gregtech.common.blocks.BlockTurbineCasing.TurbineCasingType;
-import gregtech.common.blocks.MetaBlocks;
-import gregtech.common.metatileentities.multi.multiblockpart.MetaTileEntityRotorHolder;
 import net.minecraft.block.state.IBlockState;
 import net.minecraft.client.resources.I18n;
 import net.minecraft.util.ResourceLocation;
@@ -31,53 +28,74 @@ import net.minecraftforge.fluids.capability.IFluidHandler;
 import javax.annotation.Nonnull;
 import java.util.List;
 
-import static gregtech.api.metatileentity.multiblock.MultiblockAbility.ABILITY_ROTOR_HOLDER;
+public class MetaTileEntityLargeTurbine extends FuelMultiblockController implements ITieredMetaTileEntity {
 
-public class MetaTileEntityLargeTurbine extends RotorHolderMultiblockController {
+    public final int tier;
+
+    public final IBlockState casingState;
+    public final IBlockState gearboxState;
+    public final ICubeRenderer casingRenderer;
+    public final boolean hasMufflerHatch;
+    public final ICubeRenderer frontOverlay;
 
     private static final int MIN_DURABILITY_TO_WARN = 10;
 
-    public enum TurbineType {
-
-        STEAM(RecipeMaps.STEAM_TURBINE_FUELS, MetaBlocks.TURBINE_CASING.getState(TurbineCasingType.STEEL_TURBINE_CASING), MetaBlocks.TURBINE_CASING.getState(TurbineCasingType.STEEL_GEARBOX), Textures.SOLID_STEEL_CASING, true, false, Textures.LARGE_STEAM_TURBINE_OVERLAY),
-        GAS(RecipeMaps.GAS_TURBINE_FUELS, MetaBlocks.TURBINE_CASING.getState(TurbineCasingType.STAINLESS_TURBINE_CASING), MetaBlocks.TURBINE_CASING.getState(TurbineCasingType.STEEL_GEARBOX), Textures.CLEAN_STAINLESS_STEEL_CASING, false, true, Textures.LARGE_GAS_TURBINE_OVERLAY),
-        PLASMA(RecipeMaps.PLASMA_GENERATOR_FUELS, MetaBlocks.TURBINE_CASING.getState(TurbineCasingType.TUNGSTENSTEEL_TURBINE_CASING), MetaBlocks.TURBINE_CASING.getState(TurbineCasingType.STEEL_GEARBOX), Textures.ROBUST_TUNGSTENSTEEL_CASING, true, false, Textures.LARGE_PLASMA_TURBINE_OVERLAY);
-
-        public final FuelRecipeMap recipeMap;
-        public final IBlockState casingState;
-        public final IBlockState gearboxState;
-        public final ICubeRenderer casingRenderer;
-        public final boolean hasOutputHatch;
-        public final boolean hasMufflerHatch;
-        public final ICubeRenderer frontOverlay;
-
-        TurbineType(FuelRecipeMap recipeMap, IBlockState casingState, IBlockState gearboxState, ICubeRenderer casingRenderer, boolean hasOutputHatch, boolean hasMufflerHatch, ICubeRenderer frontOverlay) {
-            this.recipeMap = recipeMap;
-            this.casingState = casingState;
-            this.gearboxState = gearboxState;
-            this.casingRenderer = casingRenderer;
-            this.hasOutputHatch = hasOutputHatch;
-            this.hasMufflerHatch = hasMufflerHatch;
-            this.frontOverlay = frontOverlay;
-        }
-    }
-
-    public final TurbineType turbineType;
     public IFluidHandler exportFluidHandler;
 
-    public MetaTileEntityLargeTurbine(ResourceLocation metaTileEntityId, TurbineType turbineType) {
-        super(metaTileEntityId, turbineType.recipeMap, GTValues.V[4]);
-        this.turbineType = turbineType;
-    }
-
-    @Override
-    protected FuelRecipeLogic createWorkable(long maxVoltage) {
-        return new LargeTurbineWorkableHandler(this, recipeMap, () -> energyContainer, () -> importFluidHandler);
+    public MetaTileEntityLargeTurbine(ResourceLocation metaTileEntityId, RecipeMap<?> recipeMap, int tier, IBlockState casingState, IBlockState gearboxState, ICubeRenderer casingRenderer, boolean hasMufflerHatch, ICubeRenderer frontOverlay) {
+        super(metaTileEntityId, recipeMap, tier);
+        this.casingState = casingState;
+        this.gearboxState = gearboxState;
+        this.casingRenderer = casingRenderer;
+        this.hasMufflerHatch = hasMufflerHatch;
+        this.frontOverlay = frontOverlay;
+        this.tier = tier;
+        this.recipeMapWorkable = new LargeTurbineWorkableHandler(this, tier);
+        this.recipeMapWorkable.enableOverclockVoltage();
+        this.recipeMapWorkable.setOverclockTier(tier);
     }
 
     @Override
     public MetaTileEntity createMetaTileEntity(MetaTileEntityHolder holder) {
-        return new MetaTileEntityLargeTurbine(metaTileEntityId, turbineType);
+        return new MetaTileEntityLargeTurbine(metaTileEntityId, recipeMap, tier, casingState, gearboxState, casingRenderer, hasMufflerHatch, frontOverlay);
+    }
+
+    public IRotorHolder getRotorHolder() {
+        List<IRotorHolder> abilities = getAbilities(MultiblockAbility.ROTOR_HOLDER);
+        if (abilities.isEmpty())
+            return null;
+        return abilities.get(0);
+    }
+
+//    @Override
+//    protected void updateFormedValid() {
+//        if (isRotorFaceFree()) {
+//            super.updateFormedValid();
+//        }
+//    }
+
+    @Override
+    public void invalidateStructure() {
+        this.exportFluidHandler = null;
+        super.invalidateStructure();
+    }
+
+    /**
+     * @return true if turbine is formed and it's face is free and contains
+     * only air blocks in front of rotor holder
+     */
+    public boolean isRotorFaceFree() {
+        IRotorHolder rotorHolder = getRotorHolder();
+        if (rotorHolder != null)
+            return isStructureFormed() && getRotorHolder().isFrontFaceFree();
+        return false;
+    }
+
+    /**
+     * @return true if structure formed, workable is active and front face is free
+     */
+    public boolean isActive() {
+        return recipeMapWorkable.isWorkingEnabled() && recipeMapWorkable.isActive() && ((LargeTurbineWorkableHandler) recipeMapWorkable).shouldRun();
     }
 
     @Override
@@ -87,67 +105,61 @@ public class MetaTileEntityLargeTurbine extends RotorHolderMultiblockController 
     }
 
     @Override
-    public void invalidateStructure() {
-        super.invalidateStructure();
-        this.exportFluidHandler = null;
-    }
-
-    @Override
-    public int getRotorSpeedIncrement() {
-        return 1;
-    }
-
-    @Override
-    public int getRotorSpeedDecrement() {
-        return -3;
-    }
-
-    @Override
     protected void addDisplayText(List<ITextComponent> textList) {
-//        if (isStructureFormed()) {
-//            MetaTileEntityRotorHolder rotorHolder = getRotorHolder();
-//            FluidStack fuelStack = ((LargeTurbineWorkableHandler) workableHandler).getFuelStack();
-//            int fuelAmount = fuelStack == null ? 0 : fuelStack.amount;
-//
-//            ITextComponent fuelName = new TextComponentTranslation(fuelAmount == 0 ? "gregtech.fluid.empty" : fuelStack.getUnlocalizedName());
-//            textList.add(new TextComponentTranslation("gregtech.multiblock.turbine.fuel_amount", fuelAmount, fuelName));
-//
-//            if (rotorHolder.getRotorEfficiency() > 0.0) {
-//                textList.add(new TextComponentTranslation("gregtech.multiblock.turbine.rotor_speed", rotorHolder.getCurrentRotorSpeed(), rotorHolder.getMaxRotorSpeed()));
-//                textList.add(new TextComponentTranslation("gregtech.multiblock.turbine.rotor_efficiency", (int) (rotorHolder.getRotorEfficiency() * 100)));
-//                int rotorDurability = (int) (rotorHolder.getRotorDurability() * 100);
-//                if (rotorDurability > MIN_DURABILITY_TO_WARN) {
-//                    textList.add(new TextComponentTranslation("gregtech.multiblock.turbine.rotor_durability", rotorDurability));
-//                } else {
-//                    textList.add(new TextComponentTranslation("gregtech.multiblock.turbine.low_rotor_durability",
-//                            MIN_DURABILITY_TO_WARN, rotorDurability).setStyle(new Style().setColor(TextFormatting.RED)));
-//                }
-//            }
-//            if(!isRotorFaceFree()) {
-//                textList.add(new TextComponentTranslation("gregtech.multiblock.turbine.obstructed")
-//                        .setStyle(new Style().setColor(TextFormatting.RED)));
-//            }
-//        }
+        if (isStructureFormed()) {
+            IRotorHolder rotorHolder = getRotorHolder();
+            FluidStack fuelStack = ((LargeTurbineWorkableHandler) recipeMapWorkable).getInputFluidStack();
+            int fuelAmount = fuelStack == null ? 0 : fuelStack.amount;
+
+            ITextComponent fuelName = new TextComponentTranslation(fuelAmount == 0 ? "gregtech.fluid.empty" : fuelStack.getUnlocalizedName());
+            textList.add(new TextComponentTranslation("gregtech.multiblock.turbine.fuel_amount", fuelAmount, fuelName));
+
+            if (rotorHolder.getRotorEfficiency() > 0) {
+                textList.add(new TextComponentTranslation("gregtech.multiblock.turbine.rotor_speed", rotorHolder.getRotorSpeed(), rotorHolder.getMaxRotorHolderSpeed()));
+                textList.add(new TextComponentTranslation("gregtech.multiblock.turbine.efficiency", rotorHolder.getTotalEfficiency()));
+
+                long maxProduction = ((LargeTurbineWorkableHandler) recipeMapWorkable).getMaxVoltage();
+                long currentProduction = isActive() ? ((LargeTurbineWorkableHandler) recipeMapWorkable).boostProduction((int) maxProduction) : 0;
+                if (currentProduction >= maxProduction) {
+                    textList.add(new TextComponentTranslation("gregtech.multiblock.turbine.energy_per_tick_maxed", maxProduction));
+                } else {
+                    textList.add(new TextComponentTranslation("gregtech.multiblock.turbine.energy_per_tick", currentProduction, maxProduction));
+                }
+
+                int rotorDurability = rotorHolder.getRotorDurabilityPercent();
+                if (rotorDurability > MIN_DURABILITY_TO_WARN) {
+                    textList.add(new TextComponentTranslation("gregtech.multiblock.turbine.rotor_durability", rotorDurability));
+                } else {
+                    textList.add(new TextComponentTranslation("gregtech.multiblock.turbine.rotor_durability", rotorDurability).setStyle(new Style().setColor(TextFormatting.RED)));
+                }
+            }
+            if(!isRotorFaceFree()) {
+                textList.add(new TextComponentTranslation("gregtech.multiblock.turbine.obstructed")
+                        .setStyle(new Style().setColor(TextFormatting.RED)));
+            }
+        }
         super.addDisplayText(textList);
     }
 
     @Override
     protected BlockPattern createStructurePattern() {
-        if (turbineType == null)
-            return null;
-
-        TraceabilityPredicate predicate = states(getCasingState()).or(autoAbilities(true, true, true, true, false));
+        TraceabilityPredicate holderPredicate = new TraceabilityPredicate();
         return FactoryBlockPattern.start()
                 .aisle("CCCC", "CHHC", "CCCC")
-                .aisle("CHHC", "RGGD", "CTTC")
+                .aisle("CHHC", "RGGR", "CTTC")
                 .aisle("CCCC", "CSHC", "CCCC")
                 .where('S', selfPredicate())
                 .where('G', states(getGearBoxState()))
                 .where('C', states(getCasingState()))
-                .where('R', any())
-                .where('D', abilities(MultiblockAbility.OUTPUT_ENERGY))
-                .where('H', predicate)
-                .where('T', predicate.or(autoAbilities(false, false, false, false, true)))
+                .where('R', metaTileEntities(MultiblockAbility.REGISTRY.get(MultiblockAbility.ROTOR_HOLDER).stream()
+                        .filter(mte -> (mte instanceof ITieredMetaTileEntity) && (((ITieredMetaTileEntity) mte).getTier() >= tier))
+                        .toArray(MetaTileEntity[]::new))
+                        .addTooltips("gregtech.multiblock.pattern.clear_amount_3")
+                        .addTooltips(I18n.format("gregtech.multiblock.pattern.error.limited.1", GTValues.VN[tier]))
+                        .setExactLimit(1)
+                        .or(abilities(MultiblockAbility.OUTPUT_ENERGY)).setExactLimit(1))
+                .where('H', states(getCasingState()).or(autoAbilities(false, true, false, false, true, true, false)))
+                .where('T', states(getCasingState()).or(autoAbilities(false, true)))
                 .build();
     }
 
@@ -159,35 +171,39 @@ public class MetaTileEntityLargeTurbine extends RotorHolderMultiblockController 
     @Override
     public boolean canShare() {
         return false;
-
     }
 
     public IBlockState getCasingState() {
-        return turbineType.casingState;
+        return casingState;
     }
 
     public IBlockState getGearBoxState() {
-        return turbineType.gearboxState;
+        return gearboxState;
     }
 
     @Override
     public ICubeRenderer getBaseTexture(IMultiblockPart sourcePart) {
-        return turbineType.casingRenderer;
+        return casingRenderer;
     }
 
     @Nonnull
     @Override
     protected ICubeRenderer getFrontOverlay() {
-        return turbineType.frontOverlay;
+        return frontOverlay;
     }
 
     @Override
     public boolean hasMufflerMechanics() {
-        return turbineType.hasMufflerHatch;
+        return hasMufflerHatch;
     }
 
     @Override
     public boolean isStructureObstructed() {
         return !isRotorFaceFree();
+    }
+
+    @Override
+    public int getTier() {
+        return tier;
     }
 }
