+++ src/main/java/gregtech/common/metatileentities/multi/multiblockpart/MetaTileEntityRotorHolder.java
@@ -4,18 +4,19 @@ import codechicken.lib.raytracer.CuboidRayTraceResult;
 import codechicken.lib.render.CCRenderState;
 import codechicken.lib.render.pipeline.IVertexOperation;
 import codechicken.lib.vec.Matrix4;
+import gregtech.api.capability.IRotorHolder;
 import gregtech.api.damagesources.DamageSources;
 import gregtech.api.gui.GuiTextures;
 import gregtech.api.gui.ModularUI;
+import gregtech.api.metatileentity.ITieredMetaTileEntity;
 import gregtech.api.metatileentity.MetaTileEntity;
 import gregtech.api.metatileentity.MetaTileEntityHolder;
 import gregtech.api.metatileentity.multiblock.IMultiblockAbilityPart;
 import gregtech.api.metatileentity.multiblock.MultiblockAbility;
 import gregtech.client.renderer.texture.Textures;
-import gregtech.api.unification.material.Material;
 import gregtech.common.advancement.GTTriggers;
 import gregtech.common.items.behaviors.TurbineRotorBehavior;
-import gregtech.common.metatileentities.multi.electric.generator.RotorHolderMultiblockController;
+import gregtech.common.metatileentities.multi.electric.generator.MetaTileEntityLargeTurbine;
 import net.minecraft.block.state.IBlockState;
 import net.minecraft.client.resources.I18n;
 import net.minecraft.entity.player.EntityPlayer;
@@ -24,257 +25,48 @@ import net.minecraft.item.ItemStack;
 import net.minecraft.nbt.NBTTagCompound;
 import net.minecraft.network.PacketBuffer;
 import net.minecraft.util.EnumFacing;
-import net.minecraft.util.EnumFacing.Axis;
 import net.minecraft.util.EnumHand;
 import net.minecraft.util.NonNullList;
 import net.minecraft.util.ResourceLocation;
 import net.minecraft.util.math.BlockPos;
-import net.minecraft.util.math.MathHelper;
 import net.minecraft.world.World;
+import net.minecraftforge.fml.relauncher.Side;
+import net.minecraftforge.fml.relauncher.SideOnly;
 import net.minecraftforge.items.ItemStackHandler;
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 import java.util.List;
 
-import static gregtech.api.capability.GregtechDataCodes.IS_ROTOR_LOOPING;
-import static gregtech.api.capability.GregtechDataCodes.UPDATE_ROTOR_COLOR;
+public class MetaTileEntityRotorHolder extends MetaTileEntityMultiblockPart implements IMultiblockAbilityPart<IRotorHolder>, IRotorHolder {
 
-public class MetaTileEntityRotorHolder extends MetaTileEntityMultiblockPart implements IMultiblockAbilityPart<MetaTileEntityRotorHolder> {
+    static final int SPEED_INCREMENT = 1;
+    static final int SPEED_DECREMENT = 3;
 
-    private static final int NORMAL_MAXIMUM_SPEED = 6000;
-    private static final float DAMAGE_PER_INTERACT = 40.0f;
+    private final InventoryRotorHolder inventory;
 
-    private final InventoryRotorHolder rotorInventory;
-    private final int maxRotorSpeed;
-    private int currentRotorSpeed;
-
-    private boolean isRotorLooping;
+    private final int maxSpeed;
+    private int currentSpeed;
     private int rotorColor = -1;
+    private boolean isRotorSpinning;
     private boolean frontFaceFree;
 
-    public MetaTileEntityRotorHolder(ResourceLocation metaTileEntityId, int tier, int maxSpeed) {
+    public MetaTileEntityRotorHolder(ResourceLocation metaTileEntityId, int tier) {
         super(metaTileEntityId, tier);
-        this.maxRotorSpeed = maxSpeed;
-        this.currentRotorSpeed = 0;
-        this.rotorInventory = new InventoryRotorHolder();
-    }
-
-    public MetaTileEntityRotorHolder(ResourceLocation metaTileEntityId, int tier, float speedMultiplier) {
-        this(metaTileEntityId, tier, (int) (NORMAL_MAXIMUM_SPEED * speedMultiplier));
-    }
-
-    public ItemStackHandler getRotorInventory() {
-        return rotorInventory;
+        this.inventory = new InventoryRotorHolder();
+        this.maxSpeed = 2000 + 1000 * tier;
     }
 
     @Override
     public MetaTileEntity createMetaTileEntity(MetaTileEntityHolder holder) {
-        return new MetaTileEntityRotorHolder(metaTileEntityId, getTier(), maxRotorSpeed);
+        return new MetaTileEntityRotorHolder(metaTileEntityId, getTier());
     }
 
     @Override
-    public void update() {
-        super.update();
-
-        if (getWorld().isRemote) {
-            return;
-        }
-        if (getOffsetTimer() % 10 == 0) {
-            this.frontFaceFree = checkTurbineFaceFree();
-        }
-
-        RotorHolderMultiblockController controller = (RotorHolderMultiblockController) getController();
-
-        if (!isHasRotor()) {
-            resetRotorSpeed();
-        } else if (controller != null) {
-            boolean isControllerActive = controller.isActive();
-
-            if (isControllerActive && currentRotorSpeed < maxRotorSpeed) {
-                incrementSpeed(controller.getRotorSpeedIncrement());
-            } else if (!isControllerActive && currentRotorSpeed > 0) {
-                incrementSpeed(controller.getRotorSpeedDecrement());
-            }
-        }
-    }
-
-    /**
-     * @return true if front face is free and contains only air blocks in 3x3 area
-     */
-    public boolean isFrontFaceFree() {
-        return frontFaceFree;
-    }
-
-    private boolean checkTurbineFaceFree() {
-        EnumFacing facing = getFrontFacing();
-        //TODO, this also needs to check Y for freedom wrench implementation
-        boolean permuteXZ = facing.getAxis() == Axis.Z;
-        BlockPos centerPos = getPos().offset(facing);
-        for (int x = -1; x < 2; x++) {
-            for (int y = -1; y < 2; y++) {
-                BlockPos blockPos = centerPos.add(permuteXZ ? x : 0, y, permuteXZ ? 0 : x);
-                IBlockState blockState = getWorld().getBlockState(blockPos);
-                if (!blockState.getBlock().isAir(blockState, getWorld(), blockPos)) {
-                    return false;
-                }
-            }
-        }
-        return true;
-    }
-
-    private boolean onRotorHolderInteract(EntityPlayer player) {
-        if (player.capabilities.isCreativeMode) {
-            return false;
-        }
-        if (!getWorld().isRemote && isRotorLooping) {
-            double relativeSpeed = getRelativeRotorSpeed();
-            float damageApplied = (float) (DAMAGE_PER_INTERACT * relativeSpeed);
-            player.attackEntityFrom(DamageSources.getTurbineDamage(), damageApplied);
-            GTTriggers.ROTOR_HOLDER_DEATH.trigger((EntityPlayerMP) player);
-            return true;
-        }
-        return isRotorLooping;
-    }
-
-    /**
-     * @return current rotor speed, relative to NORMAL_MAXIMUM_SPEED
-     * used for scaling stats like produced EU, rotor damage and interact damage
-     */
-    public double getRelativeRotorSpeed() {
-        return currentRotorSpeed == 0 ? 0.01 : currentRotorSpeed / (NORMAL_MAXIMUM_SPEED * 1.0);
-    }
-
-    /**
-     * @return true if current rotor is still looping, i.e it's current speed > 0
-     * this can return true even if multiblock is not formed
-     */
-    public boolean isRotorLooping() {
-        return isRotorLooping;
-    }
-
-    /**
-     * @return true if this holder has rotor in it's inventory
-     * this will return true even if it's client side and used for rendering
-     */
-    public boolean isHasRotor() {
-        return rotorColor != -1;
-    }
-
-    public void resetRotorSpeed() {
-        currentRotorSpeed = 0;
-        markDirty();
-    }
-
-    public int getRotorColor() {
-        return rotorColor;
-    }
-
-    public double getRotorEfficiency() {
-        return rotorInventory.getRotorEfficiency();
-    }
-
-    public boolean hasRotorInInventory() {
-        return !rotorInventory.getStackInSlot(0).isEmpty();
-    }
-
-    public boolean applyDamageToRotor(int damageAmount, boolean simulate) {
-        return rotorInventory.applyDamageToRotor(damageAmount, simulate);
-    }
-
-    public double getRotorDurability() {
-        return 1.0 - rotorInventory.getRotorDamagePercentage();
-    }
-
-    public int getMaxRotorSpeed() {
-        return maxRotorSpeed;
-    }
-
-    public int getCurrentRotorSpeed() {
-        return currentRotorSpeed;
-    }
-
-    protected void incrementSpeed(int incrementSpeed) {
-        boolean lastIsLooping = currentRotorSpeed > 0;
-        this.currentRotorSpeed = MathHelper.clamp(currentRotorSpeed + incrementSpeed, 0, maxRotorSpeed);
-        this.isRotorLooping = currentRotorSpeed > 0;
-        if (isRotorLooping != lastIsLooping && !getWorld().isRemote) {
-            writeCustomData(IS_ROTOR_LOOPING, writer -> writer.writeBoolean(isRotorLooping));
-            markDirty();
-        }
-    }
-
-    private void setRotorColor(int hasRotor1) {
-        int lastHasRotor = rotorColor;
-        this.rotorColor = hasRotor1;
-        if (rotorColor != lastHasRotor && (getWorld() != null && !getWorld().isRemote)) {
-            writeCustomData(UPDATE_ROTOR_COLOR, writer -> writer.writeInt(rotorColor));
-            markDirty();
-        }
-    }
-
-    @Override
-    public void receiveCustomData(int dataId, PacketBuffer buf) {
-        super.receiveCustomData(dataId, buf);
-        if (dataId == IS_ROTOR_LOOPING) {
-            this.isRotorLooping = buf.readBoolean();
-            getHolder().scheduleChunkForRenderUpdate();
-        } else if (dataId == UPDATE_ROTOR_COLOR) {
-            this.rotorColor = buf.readInt();
-            getHolder().scheduleChunkForRenderUpdate();
-        }
-    }
-
-    @Override
-    public void writeInitialSyncData(PacketBuffer buf) {
-        super.writeInitialSyncData(buf);
-        buf.writeBoolean(isRotorLooping);
-        buf.writeInt(rotorColor);
-    }
-
-    @Override
-    public void receiveInitialSyncData(PacketBuffer buf) {
-        super.receiveInitialSyncData(buf);
-        this.isRotorLooping = buf.readBoolean();
-        this.rotorColor = buf.readInt();
-        getHolder().scheduleChunkForRenderUpdate();
-    }
-
-    @Override
-    public void clearMachineInventory(NonNullList<ItemStack> itemBuffer) {
-        super.clearMachineInventory(itemBuffer);
-        clearInventory(itemBuffer, rotorInventory);
-    }
-
-    @Override
-    public NBTTagCompound writeToNBT(NBTTagCompound data) {
-        super.writeToNBT(data);
-        data.setTag("RotorInventory", rotorInventory.serializeNBT());
-        data.setInteger("CurrentSpeed", currentRotorSpeed);
-        return data;
-    }
-
-    @Override
-    public void readFromNBT(NBTTagCompound data) {
-        super.readFromNBT(data);
-        this.rotorInventory.deserializeNBT(data.getCompoundTag("RotorInventory"));
-        this.currentRotorSpeed = data.getInteger("CurrentSpeed");
-        this.isRotorLooping = currentRotorSpeed > 0;
-    }
-
-    @Override
-    public void renderMetaTileEntity(CCRenderState renderState, Matrix4 translation, IVertexOperation[] pipeline) {
-        super.renderMetaTileEntity(renderState, translation, pipeline);
-        Textures.ROTOR_HOLDER_OVERLAY.renderSided(getFrontFacing(), renderState, translation, pipeline);
-        Textures.LARGE_TURBINE_ROTOR_RENDERER.renderSided(renderState, translation, pipeline, getFrontFacing(),
-                getController() != null, isHasRotor(), isRotorLooping(), rotorColor);
-    }
-
-    @Override
-    protected ModularUI createUI(EntityPlayer entityPlayer) {
+    protected ModularUI createUI(@Nonnull EntityPlayer entityPlayer) {
         return ModularUI.defaultBuilder()
                 .label(6, 6, getMetaFullName())
-                .slot(rotorInventory, 0, 79, 36, GuiTextures.SLOT, GuiTextures.TURBINE_OVERLAY)
+                .slot(inventory, 0, 79, 36, GuiTextures.SLOT, GuiTextures.TURBINE_OVERLAY)
                 .bindPlayerInventory(entityPlayer.inventory)
                 .build(getHolder(), entityPlayer);
     }
@@ -288,15 +80,200 @@ public class MetaTileEntityRotorHolder extends MetaTileEntityMultiblockPart impl
     }
 
     @Override
+    public MultiblockAbility<IRotorHolder> getAbility() {
+        return MultiblockAbility.ROTOR_HOLDER;
+    }
+
+    @Override
+    public void update() {
+        super.update();
+
+        if (getWorld().isRemote) {
+            return;
+        }
+        if (getOffsetTimer() % 10 == 0) {
+            this.frontFaceFree = checkTurbineFaceFree();
+            if (!frontFaceFree || !hasRotor()) {
+                setCurrentSpeed(0);
+            }
+        }
+
+        MetaTileEntityLargeTurbine controller = (MetaTileEntityLargeTurbine) getController();
+
+        if (controller != null && controller.isActive()) {
+            if (currentSpeed < maxSpeed) {
+                setCurrentSpeed(currentSpeed + SPEED_INCREMENT);
+            }
+            if (getOffsetTimer() % 20 == 0)
+                damageRotor(1 + controller.getNumMaintenanceProblems());
+        } else if (currentSpeed > 0) {
+            setCurrentSpeed(Math.max(0, currentSpeed - SPEED_DECREMENT));
+        }
+    }
+
+    void setCurrentSpeed(int speed) {
+        if (currentSpeed != speed) {
+            currentSpeed = speed;
+            setRotorSpinning(currentSpeed > 0);
+            markDirty();
+        }
+    }
+
+    void setRotorSpinning(boolean spinning) {
+        if (isRotorSpinning != spinning) {
+            isRotorSpinning = spinning;
+            writeCustomData(42069, buf -> buf.writeBoolean(isRotorSpinning));
+        }
+    }
+
+    @Override
+    public void registerAbilities(@Nonnull List<IRotorHolder> abilityList) {
+        abilityList.add(this);
+    }
+
+    @Override
+    public boolean canPartShare() {
+        return false;
+    }
+
+    /**
+     * @return true if front face is free and contains only air blocks in 3x3 area
+     */
+    public boolean isFrontFaceFree() {
+        return frontFaceFree;
+    }
+
+    private boolean checkTurbineFaceFree() {
+        EnumFacing facing = getFrontFacing();
+        //TODO, this also needs to check Y for freedom wrench implementation
+        boolean permuteXZ = facing.getAxis() == EnumFacing.Axis.Z;
+        BlockPos centerPos = getPos().offset(facing);
+        for (int x = -1; x < 2; x++) {
+            for (int y = -1; y < 2; y++) {
+                BlockPos blockPos = centerPos.add(permuteXZ ? x : 0, y, permuteXZ ? 0 : x);
+                IBlockState blockState = getWorld().getBlockState(blockPos);
+                if (!blockState.getBlock().isAir(blockState, getWorld(), blockPos)) {
+                    return false;
+                }
+            }
+        }
+        return true;
+    }
+
+    private boolean onRotorHolderInteract(@Nonnull EntityPlayer player) {
+        if (player.isCreative())
+            return false;
+
+        if (!getWorld().isRemote && isRotorSpinning) {
+            float damageApplied = Math.min(1, currentSpeed / 1000);
+            player.attackEntityFrom(DamageSources.getTurbineDamage(), damageApplied);
+            GTTriggers.ROTOR_HOLDER_DEATH.trigger((EntityPlayerMP) player);
+            return true;
+        }
+        return isRotorSpinning;
+    }
+
+    /**
+     * returns true on both the Client and Server
+     *
+     * @return whether there is a rotor in the holder
+     */
+    @Override
+    public boolean hasRotor() {
+        return rotorColor != -1;
+    }
+
+    /**
+     * returns true on only the Server
+     *
+     * @return whether there is a rotor in the holder
+     */
+    @SideOnly(Side.SERVER)
+    @Override
+    public boolean hasRotorServer() {
+        return inventory.hasRotor();
+    }
+
+    protected void setRotorColor(int color) {
+        this.rotorColor = color;
+    }
+
+    protected int getRotorColor() {
+        return rotorColor;
+    }
+
+    @Override
+    public boolean isRotorSpinning() {
+        return isRotorSpinning;
+    }
+
+    @Override
+    public int getRotorSpeed() {
+        return this.currentSpeed;
+    }
+
+    @Override
+    public int getRotorEfficiency() {
+        return inventory.getRotorEfficiency();
+    }
+
+    @Override
+    public int getRotorPower() {
+        return inventory.getRotorPower();
+    }
+
+    @Override
+    public int getRotorDurabilityPercent() {
+        return inventory.getRotorDurabilityPercent();
+    }
+
+    @Override
+    public void damageRotor(int amount) {
+        inventory.damageRotor(amount);
+    }
+
+    @Override
+    public int getMaxRotorHolderSpeed() {
+        return this.maxSpeed;
+    }
+
+    /**
+     * calculates the holder's power multiplier: 2x per tier above the multiblock controller
+     *
+     * @return the power multiplier provided by the rotor holder
+     */
+    @Override
+    public int getHolderPowerMultiplier() {
+        int tierDifference = getTierDifference();
+        if (tierDifference == -1)
+            return -1;
+
+        return (int) Math.pow(2, getTierDifference());
+    }
+
+    @Override
+    public int getHolderEfficiency() {
+        int tierDifference = getTierDifference();
+        if (tierDifference == -1)
+            return -1;
+
+        return 100 + 10 * tierDifference;
+    }
+
+    private int getTierDifference() {
+        if (getController() instanceof ITieredMetaTileEntity)
+            return getTier() - ((ITieredMetaTileEntity) getController()).getTier();
+        return -1;
+    }
+
+    @Override
     public boolean onRightClick(EntityPlayer playerIn, EnumHand hand, EnumFacing facing, CuboidRayTraceResult hitResult) {
-        return onRotorHolderInteract(playerIn) ||
-                super.onRightClick(playerIn, hand, facing, hitResult);
+        return onRotorHolderInteract(playerIn) || super.onRightClick(playerIn, hand, facing, hitResult);
     }
 
     @Override
     public boolean onWrenchClick(EntityPlayer playerIn, EnumHand hand, EnumFacing facing, CuboidRayTraceResult hitResult) {
-        return onRotorHolderInteract(playerIn) ||
-                super.onWrenchClick(playerIn, hand, facing, hitResult);
+        return onRotorHolderInteract(playerIn) || super.onWrenchClick(playerIn, hand, facing, hitResult);
     }
 
     @Override
@@ -310,18 +287,56 @@ public class MetaTileEntityRotorHolder extends MetaTileEntityMultiblockPart impl
     }
 
     @Override
-    public MultiblockAbility<MetaTileEntityRotorHolder> getAbility() {
-        return MultiblockAbility.ABILITY_ROTOR_HOLDER;
+    public void clearMachineInventory(NonNullList<ItemStack> itemBuffer) {
+        super.clearMachineInventory(itemBuffer);
+        clearInventory(itemBuffer, inventory);
     }
 
     @Override
-    public void registerAbilities(List<MetaTileEntityRotorHolder> abilityList) {
-        abilityList.add(this);
+    public NBTTagCompound writeToNBT(NBTTagCompound data) {
+        super.writeToNBT(data);
+        data.setTag("inventory", inventory.serializeNBT());
+        data.setInteger("currentSpeed", currentSpeed);
+        return data;
     }
 
     @Override
-    public boolean canPartShare() {
-        return false;
+    public void readFromNBT(NBTTagCompound data) {
+        super.readFromNBT(data);
+        this.inventory.deserializeNBT(data.getCompoundTag("inventory"));
+        this.currentSpeed = data.getInteger("currentSpeed");
+    }
+
+    @Override
+    public void receiveCustomData(int dataId, PacketBuffer buf) {
+        super.receiveCustomData(dataId, buf);
+        if (dataId == 42069) {
+            this.isRotorSpinning = buf.readBoolean();
+            scheduleRenderUpdate();
+        }
+    }
+
+    @Override
+    public void writeInitialSyncData(PacketBuffer buf) {
+        super.writeInitialSyncData(buf);
+        buf.writeBoolean(isRotorSpinning);
+        buf.writeInt(rotorColor);
+    }
+
+    @Override
+    public void receiveInitialSyncData(PacketBuffer buf) {
+        super.receiveInitialSyncData(buf);
+        this.isRotorSpinning = buf.readBoolean();
+        this.rotorColor = buf.readInt();
+        getHolder().scheduleChunkForRenderUpdate();
+    }
+
+    @Override
+    public void renderMetaTileEntity(CCRenderState renderState, Matrix4 translation, IVertexOperation[] pipeline) {
+        super.renderMetaTileEntity(renderState, translation, pipeline);
+        Textures.ROTOR_HOLDER_OVERLAY.renderSided(getFrontFacing(), renderState, translation, pipeline);
+        Textures.LARGE_TURBINE_ROTOR_RENDERER.renderSided(renderState, translation, pipeline, getFrontFacing(),
+                getController() != null, hasRotor(), isRotorSpinning, getRotorColor());
     }
 
     private class InventoryRotorHolder extends ItemStackHandler {
@@ -343,68 +358,79 @@ public class MetaTileEntityRotorHolder extends MetaTileEntityMultiblockPart impl
         @Override
         protected void onContentsChanged(int slot) {
             setRotorColor(getRotorColor());
+            scheduleRenderUpdate();
+        }
+
+        @Nullable
+        private ItemStack getTurbineStack() {
+            if (!hasRotor())
+                return null;
+            return getStackInSlot(0);
+        }
+
+        @Nullable
+        private TurbineRotorBehavior getTurbineBehavior() {
+            ItemStack stack = getStackInSlot(0);
+            if (stack.isEmpty())
+                return null;
+
+            return TurbineRotorBehavior.getInstanceFor(stack);
+        }
+
+        private boolean hasRotor() {
+            return getTurbineBehavior() != null;
         }
 
         private int getRotorColor() {
-            ItemStack itemStack = getStackInSlot(0);
-            if (itemStack.isEmpty()) {
+            if (!hasRotor())
                 return -1;
-            }
-            TurbineRotorBehavior behavior = TurbineRotorBehavior.getInstanceFor(itemStack);
-            if (behavior == null) {
+            //noinspection ConstantConditions
+            return getTurbineBehavior().getPartMaterial(getStackInSlot(0)).getMaterialRGB();
+
+        }
+
+        private int getRotorDurabilityPercent() {
+            if (!hasRotor())
+                return 0;
+            //noinspection ConstantConditions
+            return getTurbineBehavior().getRotorDurabilityPercent(getStackInSlot(0));
+        }
+
+        private int getRotorEfficiency() {
+            if (!hasRotor())
                 return -1;
-            }
-            Material material = behavior.getPartMaterial(itemStack);
-            return material.getMaterialRGB();
+
+            //noinspection ConstantConditions
+            return getTurbineBehavior().getRotorEfficiency(getTurbineStack());
         }
 
-        public double getRotorEfficiency() {
-            ItemStack itemStack = getStackInSlot(0);
-            if (itemStack.isEmpty()) {
-                return 0.0;
-            }
-            TurbineRotorBehavior turbineRotorBehavior = TurbineRotorBehavior.getInstanceFor(itemStack);
-            if (turbineRotorBehavior != null) {
-                return turbineRotorBehavior.getRotorEfficiency(itemStack);
-            }
-            return 0.0;
+        private int getRotorPower() {
+            if (!hasRotor())
+                return -1;
+
+            //noinspection ConstantConditions
+            return getTurbineBehavior().getRotorPower(getTurbineStack());
         }
 
-        public double getRotorDamagePercentage() {
-            ItemStack itemStack = getStackInSlot(0);
-            if (itemStack.isEmpty()) {
-                return 0.0;
-            }
-            TurbineRotorBehavior turbineRotorBehavior = TurbineRotorBehavior.getInstanceFor(itemStack);
-            if (turbineRotorBehavior != null) {
-                return turbineRotorBehavior.getPartDamage(itemStack) / (turbineRotorBehavior.getPartMaxDurability(itemStack) * 1.0);
-            }
-            return 0.0;
+        private void damageRotor(int damageAmount) {
+            if (!hasRotor())
+                return;
+            //noinspection ConstantConditions
+            getTurbineBehavior().applyRotorDamage(getStackInSlot(0), damageAmount);
         }
 
-        public boolean applyDamageToRotor(int damageAmount, boolean simulate) {
-            ItemStack itemStack = getStackInSlot(0);
-            if (itemStack.isEmpty()) {
-                return false;
-            }
-            TurbineRotorBehavior turbineRotorBehavior = TurbineRotorBehavior.getInstanceFor(itemStack);
-            if (turbineRotorBehavior != null) {
-                if (!simulate) {
-                    turbineRotorBehavior.applyRotorDamage(itemStack, damageAmount);
-                }
-                return true;
-            }
-            return false;
+        @Override
+        public boolean isItemValid(int slot, @Nonnull ItemStack stack) {
+            return TurbineRotorBehavior.getInstanceFor(stack) != null && super.isItemValid(slot, stack);
         }
 
         @Nonnull
         @Override
-        public ItemStack insertItem(int slot, @Nonnull ItemStack stack, boolean simulate) {
-            if (TurbineRotorBehavior.getInstanceFor(stack) != null) {
-                return super.insertItem(slot, stack, simulate);
-            }
-            return stack;
+        public ItemStack extractItem(int slot, int amount, boolean simulate) {
+            ItemStack itemStack = super.extractItem(slot, amount, simulate);
+            if (!simulate && itemStack != ItemStack.EMPTY)
+                setRotorColor(-1);
+            return itemStack;
         }
     }
-
 }
