+++ src/main/java/gregtech/loaders/recipe/handlers/PartsRecipeHandler.java
@@ -18,7 +18,6 @@ import gregtech.api.util.GTUtility;
 import gregtech.common.ConfigHolder;
 import gregtech.common.items.MetaItems;
 import gregtech.common.items.behaviors.TurbineRotorBehavior;
-import gregtech.common.items.behaviors.TurbineRotorBehavior2;
 import net.minecraft.item.EnumDyeColor;
 import net.minecraft.item.ItemStack;
 
@@ -433,7 +432,7 @@ public class PartsRecipeHandler {
     public static void processTurbine(OrePrefix toolPrefix, Material material, IngotProperty property) {
         ItemStack rotorStack = MetaItems.TURBINE_ROTOR.getStackForm();
 //        noinspection ConstantConditions
-        TurbineRotorBehavior2.getInstanceFor(rotorStack).setPartMaterial(rotorStack, material);
+        TurbineRotorBehavior.getInstanceFor(rotorStack).setPartMaterial(rotorStack, material);
 
         RecipeMaps.ASSEMBLER_RECIPES.recipeBuilder()
                 .input(OrePrefix.turbineBlade, material, 8)
