+++ src/main/resources/assets/gregtech/lang/en_us.lang
@@ -4063,7 +4063,6 @@ gregtech.multiblock.pattern.clear_amount_1=§6Must have a clear 1x1x1 space in f
 gregtech.multiblock.pattern.clear_amount_3=§6Must have a clear 3x3x1 space in front§r
 gregtech.multiblock.pattern.single=§6Only this block can be used§r
 gregtech.multiblock.pattern.location_end=§cVery End§r
-gregtech.multiblock.pattern.tier_minimum=§c%s tier or higher.§r
 
 gregtech.multiblock.blast_furnace.max_temperature=Max Temperature: %sK
 gregtech.multiblock.multi_furnace.heating_coil_level=Heating Coil Level: %s
